// client/src/utils/enums.ts
export enum PUBLIC_ROUTE {
  HOME = "/",
  LOGIN = "/login",
  SIGNUP = "/signup",
}

export enum PRIVATE_ROUTE {
  BOARD = "/board",
}

export enum MODAL_MODE {
  CREATE = "create",
  EDIT = "edit",
  VIEW = "view",
}

export enum STATUS {
  CREATED = "CREATED",
  INPROGRESS = "INPROGRESS",
  COMPLETED = "COMPLETED",
}

export enum SORT {
  LATEST = "latest",
  OLDEST =  "oldest"
}